[CORE] Eliminate compile warnings
1. `vasprintf` can return a negative value and `buff` may be NULL.

2. The `events` argument in `epoll_pwait` must not be NULL.

3. The size of `val.f.q` is 64 bits.

+++ src/tools/dynacache.c
@@ -875,9 +875,9 @@ int ReadDynaCache(const char* folder, const char* name, mapping_t* mapping, int 
             n += snprintf(buf+n, sizeof(buf)-n, " and %s non-canceled blocks\n", NicePrintSize(total_code));
             n += snprintf(buf+n, sizeof(buf)-n, "\tMapped at %p-%p, with %zu lock and %zu unaligned addresses",
                 (void*)file_header->map_addr, (void*)file_header->map_addr+file_header->map_len,
-                file_header->nLockAddresses, file_header->nUnalignedAddresses);
+                (size_t)file_header->nLockAddresses, (size_t)file_header->nUnalignedAddresses);
             if(total_compressed && n>0 && n<(int)sizeof(buf)) {
-                n += snprintf(buf+n, sizeof(buf)-n, "\n\tCompression: %d%% / %s compressed", 100ULL-total_compressed*100ULL/total_blocks ,NicePrintSize(total_compressed));
+                n += snprintf(buf+n, sizeof(buf)-n, "\n\tCompression: %llu%% / %s compressed", 100ULL-total_compressed*100ULL/total_blocks ,NicePrintSize(total_compressed));
                 if(total_uncompressed && n>0 && n<(int)sizeof(buf))
                     n += snprintf(buf+n, sizeof(buf)-n, ", %s uncompressed", NicePrintSize(total_uncompressed));
             }
